feat: add shared account utilities
