+++ src/jupiter/v6/accounts.rs
@@ -2,62 +2,8 @@ use borsh::{BorshDeserialize, BorshSerialize};
 use serde::{Deserialize, Serialize};
 use solana_program::pubkey::Pubkey;
 use substreams_solana::block_view::InstructionView;
-use thiserror::Error;
 
-// -----------------------------------------------------------------------------
-// Error type
-// -----------------------------------------------------------------------------
-#[derive(Debug, Error)]
-pub enum AccountsError {
-    #[error("missing required account `{name}` at index {index}")]
-    Missing { name: &'static str, index: usize },
-    #[error("invalid key length for `{name}` at index {index}: got {got}, want 32")]
-    InvalidLen { name: &'static str, index: usize, got: usize },
-}
-
-#[inline]
-fn to_pubkey(name: &'static str, index: usize, bytes: &[u8]) -> Result<Pubkey, AccountsError> {
-    let arr: [u8; 32] = bytes.try_into().map_err(|_| AccountsError::InvalidLen { name, index, got: bytes.len() })?;
-    Ok(Pubkey::new_from_array(arr))
-}
-
-// -----------------------------------------------------------------------------
-// Helper macro for required-only account structs
-// -----------------------------------------------------------------------------
-macro_rules! accounts {
-    ($name:ident, $getter:ident, { $($field:ident),+ $(,)? }) => {
-        #[derive(Debug, Clone, PartialEq, Eq, BorshSerialize, BorshDeserialize, Serialize, Deserialize)]
-        pub struct $name {
-            $(pub $field: Pubkey,)+
-        }
-
-        impl<'ix> TryFrom<&InstructionView<'ix>> for $name {
-            type Error = AccountsError;
-
-            fn try_from(ix: &InstructionView<'ix>) -> Result<Self, Self::Error> {
-                let accounts = ix.accounts();
-                let mut idx = 0;
-                $(
-                    let $field = {
-                        let name = stringify!($field);
-                        let a = accounts
-                            .get(idx)
-                            .ok_or(AccountsError::Missing { name, index: idx })?;
-                        let pk = to_pubkey(name, idx, &a.0)?;
-                        idx += 1;
-                        pk
-                    };
-                )+
-                let _ = idx;
-                Ok(Self { $($field,)+ })
-            }
-        }
-
-        pub fn $getter(ix: &InstructionView) -> Result<$name, AccountsError> {
-            $name::try_from(ix)
-        }
-    };
-}
+use crate::accounts::{self as _, AccountsError};
 
 // -----------------------------------------------------------------------------
 // Simple instructions
@@ -148,7 +94,7 @@ impl<'ix> TryFrom<&InstructionView<'ix>> for ExactOutRouteAccounts {
         let accounts = ix.accounts();
         let get_req = |index: usize, name: &'static str| -> Result<Pubkey, AccountsError> {
             let a = accounts.get(index).ok_or(AccountsError::Missing { name, index })?;
-            to_pubkey(name, index, &a.0)
+            crate::accounts::to_pubkey(name, index, &a.0)
         };
         let get_opt = |index: usize| -> Option<Pubkey> { accounts.get(index).and_then(|a| a.0.as_slice().try_into().ok()).map(Pubkey::new_from_array) };
         Ok(ExactOutRouteAccounts {
@@ -202,7 +148,7 @@ impl<'ix> TryFrom<&InstructionView<'ix>> for RouteAccounts {
         let accounts = ix.accounts();
         let get_req = |index: usize, name: &'static str| -> Result<Pubkey, AccountsError> {
             let a = accounts.get(index).ok_or(AccountsError::Missing { name, index })?;
-            to_pubkey(name, index, &a.0)
+            crate::accounts::to_pubkey(name, index, &a.0)
         };
         let get_opt = |index: usize| -> Option<Pubkey> { accounts.get(index).and_then(|a| a.0.as_slice().try_into().ok()).map(Pubkey::new_from_array) };
         Ok(RouteAccounts {
@@ -256,7 +202,7 @@ impl<'ix> TryFrom<&InstructionView<'ix>> for RouteWithTokenLedgerAccounts {
         let accounts = ix.accounts();
         let get_req = |index: usize, name: &'static str| -> Result<Pubkey, AccountsError> {
             let a = accounts.get(index).ok_or(AccountsError::Missing { name, index })?;
-            to_pubkey(name, index, &a.0)
+            crate::accounts::to_pubkey(name, index, &a.0)
         };
         let get_opt = |index: usize| -> Option<Pubkey> { accounts.get(index).and_then(|a| a.0.as_slice().try_into().ok()).map(Pubkey::new_from_array) };
         Ok(RouteWithTokenLedgerAccounts {
@@ -317,7 +263,7 @@ impl<'ix> TryFrom<&InstructionView<'ix>> for SharedAccountsExactOutRouteAccounts
         let accounts = ix.accounts();
         let get_req = |index: usize, name: &'static str| -> Result<Pubkey, AccountsError> {
             let a = accounts.get(index).ok_or(AccountsError::Missing { name, index })?;
-            to_pubkey(name, index, &a.0)
+            crate::accounts::to_pubkey(name, index, &a.0)
         };
         let get_opt = |index: usize| -> Option<Pubkey> { accounts.get(index).and_then(|a| a.0.as_slice().try_into().ok()).map(Pubkey::new_from_array) };
         Ok(SharedAccountsExactOutRouteAccounts {
@@ -367,7 +313,7 @@ impl<'ix> TryFrom<&InstructionView<'ix>> for SharedAccountsRouteAccounts {
         let accounts = ix.accounts();
         let get_req = |index: usize, name: &'static str| -> Result<Pubkey, AccountsError> {
             let a = accounts.get(index).ok_or(AccountsError::Missing { name, index })?;
-            to_pubkey(name, index, &a.0)
+            crate::accounts::to_pubkey(name, index, &a.0)
         };
         let get_opt = |index: usize| -> Option<Pubkey> { accounts.get(index).and_then(|a| a.0.as_slice().try_into().ok()).map(Pubkey::new_from_array) };
         Ok(SharedAccountsRouteAccounts {
@@ -420,7 +366,7 @@ impl<'ix> TryFrom<&InstructionView<'ix>> for SharedAccountsRouteWithTokenLedgerA
         let accounts = ix.accounts();
         let get_req = |index: usize, name: &'static str| -> Result<Pubkey, AccountsError> {
             let a = accounts.get(index).ok_or(AccountsError::Missing { name, index })?;
-            to_pubkey(name, index, &a.0)
+            crate::accounts::to_pubkey(name, index, &a.0)
         };
         let get_opt = |index: usize| -> Option<Pubkey> { accounts.get(index).and_then(|a| a.0.as_slice().try_into().ok()).map(Pubkey::new_from_array) };
         Ok(SharedAccountsRouteWithTokenLedgerAccounts {
